refactor: removed react-bootstrap dependency
