+++ package.json
@@ -33,7 +33,6 @@
     "@material-ui/icons": "^4.9.1",
     "formik": "^2.1.4",
     "react": "^16.13.0",
-    "react-bootstrap": "^1.0.0-beta.12",
     "react-dom": "^16.10.2",
     "react-router": "^5.1.2",
     "react-router-bootstrap": "^0.25.0",
