--- src/components/atoms/forms/index.ts
@@ -1,7 +1,0 @@
-import Form from './Form';
-import FormControl from './FormControl';
-import InputGroup from './InputGroup';
-import InputGroupAppend from './InputGroupAppend';
-import InputGroupText from './InputGroupText';
-
-export { Form, FormControl, InputGroup, InputGroupAppend, InputGroupText };
