--- src/components/atoms/navigation/NavBrand.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Navbar as BSNavbar } from 'react-bootstrap';
-import { LogoNavbar } from 'rifui';
-
-const NavBrand = () => (
-  <BSNavbar.Brand>
-    <LogoNavbar />
-  </BSNavbar.Brand>
-);
-
-export default NavBrand;
