+++ src/components/pages/DomainsCheckoutPage.tsx
@@ -4,11 +4,10 @@ import CombinedPriceCell from 'components/molecules/CombinedPriceCell';
 import ItemDetailRow from 'components/molecules/ItemDetailRow';
 import CheckoutPageTemplate from 'components/templates/CheckoutPageTemplate';
 import React, { useContext, useEffect } from 'react';
-import { Button } from 'react-bootstrap';
 import { useHistory } from 'react-router-dom';
 import { ROUTES } from 'routes';
 import { MARKET_ACTIONS } from 'store/Market/marketActions';
-import { Card, CardHeader, CardContent } from 'rifui/components/atoms/card';
+import { Button, Card, CardHeader, CardContent } from 'rifui';
 import MarketStore from 'store/Market/MarketStore';
 import { Web3Store } from 'rifui/providers/Web3Provider';
 import { shortenAddress } from 'rifui/utils';
